refactor(alarm_control_panel): split into a package by responsibility
alarm_control_panel.py grew to 1327 LOC mixing the platform entry point,
the household-intent panel, three single-axis sub-panels, and 900-odd
lines of base-class behaviour. Split into a package:

- __init__.py — async_setup_entry + re-exports (130 LOC)
- _base.py — BaseVerisureOwaAlarmPanel: coordinator integration,
  arm/disarm flow, force-arm context, arming-exception
  notifications (990 LOC)
- _panels.py — CombinedVerisureOwaAlarmPanel,
  _AxisSubPanelMixin, Interior/Perimeter/Annex sub-panels (275 LOC)

Two test patches need a deeper path: _notify and
get_notification_strings are now looked up in _base, so patches target
alarm_control_panel._base.<name> instead of alarm_control_panel.<name>.
async_get_current_platform stays at the package root because
async_setup_entry still lives in __init__.py.

Pure file-level reorganisation — no behaviour change. All 1378 tests
pass.
